#25734 undoing CT creation related changes
